# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_stacyvm-dev_stacyvm_internal_httputil.APIError:
    properties:
      code:
        $ref: '#/definitions/github_com_stacyvm-dev_stacyvm_internal_httputil.ErrorCode'
      message:
        type: string
    type: object
  github_com_stacyvm-dev_stacyvm_internal_httputil.ErrorCode:
    enum:
    - NOT_FOUND
    - BAD_REQUEST
    - INTERNAL_ERROR
    - UNAUTHORIZED
    - CONFLICT
    - UNAVAILABLE
    type: string
    x-enum-varnames:
    - CodeNotFound
    - CodeBadRequest
    - CodeInternal
    - CodeUnauth
    - CodeConflict
    - CodeUnavailable
...
  internal_api_routes.ProviderDetail:
    properties:
      config:
        additionalProperties:
          type: string
        type: object
      default:
        example: true
        type: boolean
      healthy:
        example: true
        type: boolean
      name:
        example: firecracker
        type: string
      sandbox_count:
        example: 3
        type: integer
    type: object
  internal_api_routes.ProviderInfo:
    properties:
      default:
        example: true
        type: boolean
      healthy:
        example: true
        type: boolean
      name:
        example: firecracker
        type: string
    type: object
...
host: localhost:7423
info:
  contact: {}
  description: StacyVM microVM sandbox orchestrator API
  title: StacyVM API
  version: "1.0"
paths:
...
  /providers:
    get:
      description: Return all registered providers with health status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_api_routes.ProviderInfo'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List providers
      tags:
      - providers
  /providers/{providerName}:
    get:
      description: Return detailed information about a specific provider
      parameters:
      - description: Provider name
        in: path
        name: providerName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_routes.ProviderDetail'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_stacyvm-dev_stacyvm_internal_httputil.APIError'
      security:
      - ApiKeyAuth: []
      summary: Get provider details
      tags:
      - providers
  /providers/test:
    post:
      description: Run health checks on all registered providers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: Test providers
      tags:
      - providers
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"